Select all the required columns ...
We need the columns in order to properly construct an entry using a model. If some columns are missing then it's possible an entry will be missing properties and will error in certain situations.

For instance, if you do ->get(['id']), then the 'collection' column would be missing, resulting in an entry without a collection.

We'll pass the user-selected columns along to the entry so that only they're used within augmented arrays.