Explicitly use `updated_at` column instead of `updated_at` within imported json data.
Because the import script was importing `updated_at` into the json data column, we will explicitly reference other SQL columns first to prevent errors with that bad data. This isn't ideal, but we shouldn't have to ask our users to manually remove `updated_at` from their JSON data, so this fixes the CP for those people.